+++ test/support/assert.js
@@ -1,5 +1,6 @@
 import * as sinon from 'sinon'
-import * as assert from 'assert'
+import assert from 'assert'
+import * as K from 'src'
 
 export function propertyValueEqual (property, expected) {
   let actual
@@ -17,4 +18,13 @@ export function propertyValueEqual (property, expected) {
 export function calledOnceWith (spy, ...args) {
   sinon.assert.calledOnce(spy)
   sinon.assert.calledWith(spy, ...args)
+}
+
+export function propertySemantics (prop) {
+  let value
+  const off = K.onValue(prop, (x) => {
+    value = x
+  })
+  assert.strictEqual(value, K.getValue(prop))
+  off()
 }