+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+## Upcoming
+
+* `K.combinePropertyObject()` now returns a proper property
+
 ## v0.5.0 2016-10-16
 
 Initial release